Rename pkg/warewulfconf to pkg/config
