+++ internal/app/wwctl/genconf/warewulfconf/print/main.go
@@ -3,7 +3,7 @@ package print
 import (
 	"fmt"
 
-	"github.com/hpcng/warewulf/internal/pkg/warewulfconf"
+	warewulfconf "github.com/hpcng/warewulf/internal/pkg/config"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v2"
 )
