+++ internal/pkg/container/mountpoints.go
@@ -3,7 +3,7 @@ package container
 import (
 	"strings"
 
-	"github.com/hpcng/warewulf/internal/pkg/warewulfconf"
+	warewulfconf "github.com/hpcng/warewulf/internal/pkg/config"
 	"github.com/hpcng/warewulf/internal/pkg/wwlog"
 )
