+++ internal/pkg/oci/defaults.go
@@ -1,7 +1,9 @@
 package oci
 
-import 	"github.com/hpcng/warewulf/internal/pkg/warewulfconf"
-import  "path/filepath"
+import (
+	warewulfconf "github.com/hpcng/warewulf/internal/pkg/config"
+	"path/filepath"
+)
 
 var defaultCachePath = filepath.Join(warewulfconf.DataStore(), "/container-cache/oci/")
