+++ internal/pkg/overlay/config.go
@@ -5,7 +5,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/hpcng/warewulf/internal/pkg/warewulfconf"
+	warewulfconf "github.com/hpcng/warewulf/internal/pkg/config"
 )
 
 func OverlaySourceTopDir() string {
