+++ internal/pkg/version/version.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/hpcng/warewulf/internal/pkg/api/routes/wwapiv1"
-	"github.com/hpcng/warewulf/internal/pkg/warewulfconf"
+	warewulfconf "github.com/hpcng/warewulf/internal/pkg/config"
 )
 
 /*
