+++ internal/pkg/warewulfd/warewulfd.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 	"syscall"
 
-	"github.com/hpcng/warewulf/internal/pkg/warewulfconf"
+	warewulfconf "github.com/hpcng/warewulf/internal/pkg/config"
 	"github.com/hpcng/warewulf/internal/pkg/wwlog"
 	"github.com/pkg/errors"
 )
